pyproject.toml
[project]
name = "ddiscord"
version = "0.3.0"
description = "Debugger for discord.py v2"
readme = "README.md"
authors = [
    { name = "FaberSid" },
    { name = "aiotter" },
]
requires-python = ">=3.13"
dependencies = [
    "discord.py>=2.6.3",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Private :: Do Not Upload",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]

[project.license]
text = "MIT"

[project.urls]
Homepage = "https://github.com/FaberSid/ddiscord"
Repository = "https://github.com/FaberSid/ddiscord"
Original = "https://github.com/aiotter/ddiscord"


[project.scripts]
ddiscord = "ddiscord:main"

[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"
